using listeners to detect the state of the add
in this version we can have the information about how the user interact with the ad, skeep it, check the advertiser page, or check why he recieved the ad and also if he checked the youtube channel of the advertiser. for the float ad, we can see either if the user checked the 'why did i reieve the add' or if he removed it.